the idiom pass: nine bindings read as their own language, and THE WIRE STOPS AT THE BINDING BOUNDARY. The maintainer's question of 2026-09-16 — are the bindings idiomatic to their language, or a naive port of one language into another (monads, effects, references, async, proxies)? — was answered by nine read-only auditors, one per language, over the sugar tier and every guest first committed on or after 2026-08-25: none is a port, all nine mostly idiomatic, no ratified spelling challenged, and the one tier that leaks is the WIRE's own encoding — its value constructors, integer vocabularies, sentinel ints, string-spelled booleans and flat spans reached guests in eight of nine bindings; the census of tracked hand-written guests naming a wire tag read haskell 51/54, ocaml 50/52, rust 10/57, csharp 10/65, java 7/52, python 5/54, go/js/swift 0. The second pattern was the compiler not asked to do the language's own checking. 65 findings (54 spelling, 1 semantics, 10 capability); the maintainer accepted every recommendation, with two rulings of his own: no rustdoc and no decision logs in comments (strip them; what a future session needs moves to the md files), and the Haskell module split is IN. THE PASS, one implementer per language: HASKELL — a KayaValue class over the wire tag (typed signals and keys; VStr gone from every guest), collectionOf @T over the Proxy, Text on the whole surface, deriving anyclass, OverloadedRecordDot, `primary` for the window literal, Build/Tpl on mtl's State, -Wall -Wcompat with the backlog fixed, getRecord, and the SPLIT: bindings/haskell/Kaya/Core.hs (Build/Tpl/Declare, signals, collections and records, KayaValue) under a KayaApp.hs that keeps everything built on top and re-exports Core, so every guest's one import is unchanged. OCAML — a phantom-typed 'a signal, typed keys, unprefixed record fields under per-type modules, Menu_role.t, the first kaya_app.mli and kaya_runtime.mli (only what a guest, a check or a gate reaches is exported) and 53 dead internals deleted behind them, not silenced. JS — window/build/pushEntry/addSection return their body's value (the `let x!: T` pattern gone from 92 sites), boolean marks, string-literal unions for appearance, sections presentation and menu roles, showAlert answering null for cancel, pickFile answering one file or null, UndoDelta with named fields, the one-shot showWindow. SWIFT — Range<Int> on runs, edits and formats, one generic KayaPrefs.get/set over a KayaPrefValue protocol, the IUO pattern in dnd, a safe date accessor, KayaSumCase deleted. C# — FileAccess, AlertChoice, NotificationOutcome and SectionsPresentation enums (KayaWire gone from every guest), nullable and implicit usings ON with the warnings fixed, TryGet and an indexer over OrderedDictionary, Tx as IDisposable, the Task-shadowing record renamed. JAVA — FileMode, AlertChoice, NotificationOutcome and SectionsPresentation enums with fromWire factories, holder classes for the array boxes in nine guests, @FunctionalInterface on 21 interfaces, one absence convention, a keyed read. GO — named types with String() for role, align, axis, appearance, symbol and menu role (a bare int no longer compiles), RecordCollection.Get, Collection un-embedded so the untyped floor stops being promoted. PYTHON — keyword-only parameters past the primary argument, IntEnum vocabularies with byte-identical refusal sentences, FileMode on PickedFile.open, pathlib for local_path, Prefs.get/set dispatching on the default's type, a KayaError base mixed into the builtins already raised, the csv module in portfolio. RUST — a PathKey trait over KayaField::from_value (nine guests' hand-rolled unwraps gone), Collection::get, #[must_use] on the three request builders that lacked it, Option and returned structs for the sentinel ids and the sixteen-tuple, Date/Time's own Display/FromStr in tasks, and the decision-log doc comments stripped from the public surface with the measured findings moved to docs/traps.md. THE GUARDS: check-sugar-surface's wire-tag clause refuses a wire constructor or constant in any hand-written guest of the nine (the generated surfaces read from gen-guests' own list, checks and benches exempt by name, a floor of forty guests per language, nine watched negatives planted 1 / blank 0; it was red by 122 findings on the tree it was written against and is green on this one); per-language prefs rows and the python keyword-only row; the notification-sentence finder reads the occurrence that names the registrar, with a decoy planted as its negative; check-abort runs the OCaml compile-fail fixtures and the Haskell KayaValue probe's four cases (a negative nobody runs is a guard nobody walks into) and its staged Haskell project carries the split; new negatives in every binding's checks file, each watched red with its count printed; go test, the Java choice and keyed fixtures, the Swift abort traps. FOUND EN ROUTE: tools/run-leg.py launched a guest bare where validate-mac wraps every leg in `timeout 120`, under which an undeclared window opens at another size (python 900x600 bare, 540x330 wrapped) — align-python failed by hand alone and three agents read it as a pre-existing red on main; the hand runner launches as the pool does now and check-gates N18 holds it (docs/traps.md). DESIGN.md's Binding conventions carry the rule; docs/deferred.md's idiom-pass entry records the survey, the rulings and the four later slices (async dialogs for Swift/C#/Java/Rust, Python row handles, the toolchain batch, the generators not taken). Validated: check-sugar-surface (the wire-tag clause at 0 findings with its nine watched negatives, seventeen runs to walk every Haskell and OCaml row through the split and the typed keys), tpl-surfaces, check-abort with the two new walls, check-gates with N18, check-python, check-doc-refs, check-ledger, check-mirror and check-file-modes green standalone; tools/validate-mac.py ALL PASS on this tree, 60/60 gates and 480 legs, align-python among them under the pool; the thirtieth matrix on this tree: mac 477, linux 775, ios 139 and android 148 legs and the sweep 60/60 ALL PASS, windows red on notes_rust alone — the recorded Windows toast class, its bundle read (a toast held the foreground 15s before type injection; the UIA walk named nothing and the toast-moment picture was skipped because the sampler's first sight fell one second before the leg, so the recorder growth is filed on the ledger) — and the windows lane ALL PASS standalone, 281 legs with the guest's 4/4 unit tests. Found and fixed on the way: check-sugar-surface's record probe wrote its doctored copy THROUGH a symlinked hop into the real Kaya/Core.hs (every shadow write is guarded now, docs/traps.md), and the OCaml dead-code deletion had cut the sugar surface no guest calls, restored and exported.

+++ bindings/go/adaptive_test.go
@@ -133,9 +133,9 @@ func TestSetAxisWritesTheAxisPropConstant(t *testing.T) {
 		t.Fatalf("SetAxis wrote widget %d prop %d source %d, want %d/%d/%d",
 			p.widget, p.prop, p.source, row, PropAxis, SourceConst)
 	}
-	if p.tag != ValueI64 || p.i64 != AxisVertical {
+	if p.tag != ValueI64 || p.i64 != int64(AxisVertical) {
 		t.Fatalf("the axis rode as tag %d value %d, want i64 AxisVertical (%d)",
-			p.tag, p.i64, AxisVertical)
+			p.tag, p.i64, int64(AxisVertical))
 	}
 }
